rename resource package
